video 16

+++ src/main/java/com/lothin/phoneshp/dto/PaginationDTO.java
@@ -7,11 +7,11 @@ import lombok.Data;
 @Builder
 public class PaginationDTO {
     private int numberOfElements;
-    private int number;
-    private int size;
-    private int totalElements;
-    private int totalPages;
-    private boolean empty;
-    private boolean first;
-    private boolean last;
+	private int number;
+	private int size;
+	private long totalElements;
+	private int totalPages;
+	private boolean empty;
+	private boolean first;
+	private boolean last;
 }
